1. optimize thread priority 2. add thread to get eps & vps status periodically 3. add interface to record override setting 4. add protocol between arm io and dose

+++ ArmCoreV1_QSPI/application/bgm_app/bgm_def.c
@@ -115,7 +115,7 @@ enum dose_fsm_state dose_fsm_state_get(enum uart_id id)
     return dose_data_info_get(id, DOSE_INFO_FSM_STATE_GET, NULL);
 }
 
-uint16_t dose_interlock_get(enum uart_id id)
+uint32_t dose_interlock_get(enum uart_id id)
 {
     return dose_data_info_get(id, DOSE_INFO_INTERLOCK_GET, NULL);
 }
